+++ packages/ui/src/views/user-management/components/dialogs/components/delete-user/types.ts
@@ -1,9 +1,6 @@
-import { IAdminListUsersStore } from '@views/user-management/types'
-
 export interface IDeleteDialogProps {
   open: boolean
   onClose: () => void
   isDeleting: boolean
   handleDeleteUser: (userUid: string) => void
-  useAdminListUsersStore: () => IAdminListUsersStore
 }
