+++ packages/ui/src/views/user-management/components/dialogs/components/edit-user/types.ts
@@ -1,4 +1,3 @@
-import { IAdminListUsersStore } from '@/views'
 import { newUserSchema } from '@views/user-management/components/dialogs/components/edit-user/schemas'
 import { z } from 'zod'
 
@@ -9,5 +8,4 @@ export interface IEditUserDialogProps {
   onClose: () => void
   handleUpdateUser: (data: { email: string; displayName: string; userID: string }) => void
   open: boolean
-  useAdminListUsersStore: () => IAdminListUsersStore
 }
