+++ packages/ui/src/views/user-management/components/dialogs/components/remove-admin/types.ts
@@ -1,9 +1,6 @@
-import { IAdminListUsersStore } from '@/views'
-
 export interface IRemoveAdminDialogProps {
   open: boolean
   onClose: () => void
   isLoading: boolean
   updateUserAdmin: (uid: string, admin: boolean) => void
-  useAdminListUsersStore: () => IAdminListUsersStore
 }
