+++ packages/ui/src/views/user-management/components/dialogs/components/reset-password/types.ts
@@ -1,8 +1,5 @@
-import { IAdminListUsersStore } from '@/views'
-
 export interface IResetPasswordDialogProps {
   onClose: () => void
   open: boolean
   handleUpdatePassword: (userId: string) => void
-  useAdminListUsersStore: () => IAdminListUsersStore
 }
