+++ packages/ui/src/views/user-management/components/dialogs/dialogs.tsx
@@ -1,11 +1,6 @@
 import { useDialogHandlers } from '@/views/user-management/components/dialogs/hooks'
-import {
-  DialogLabels,
-  IDialogErrorStates,
-  IDialogHandlers,
-  IDialogLoadingStates,
-  IUserManagementPageProps
-} from '@/views/user-management/types'
+import { DialogLabels } from '@/views/user-management/types'
+import { IDialogsProps } from '@views/user-management/components/dialogs/types'
 import { useDialogs } from '@views/user-management/providers/DialogsProvider'
 
 import { CreateUserDialog } from './components/create-user'
@@ -14,17 +9,9 @@ import { EditUserDialog } from './components/edit-user'
 import { AdminDialog } from './components/remove-admin'
 import { ResetPasswordDialog } from './components/reset-password'
 
-export const Dialogs = ({
-  useAdminListUsersStore,
-  handlers,
-  loadingStates,
-  errorStates
-}: Pick<IUserManagementPageProps, 'useAdminListUsersStore'> & {
-  handlers: IDialogHandlers
-  loadingStates: IDialogLoadingStates
-  errorStates: IDialogErrorStates
-}) => {
+export const Dialogs = ({ handlers, loadingStates, errorStates }: IDialogsProps) => {
   const { dialogsOpenState, closeDialog } = useDialogs()
+
   const { handleDeleteUser, handleUpdateUser, handleUpdatePassword, handleUpdateUserAdmin, handleCreateUser } =
     useDialogHandlers(handlers)
   const { isDeletingUser, isUpdatingUser, isUpdatingUserAdmin, isCreatingUser } = loadingStates
@@ -34,28 +21,24 @@ export const Dialogs = ({
     <>
       <DeleteUserDialog
         open={dialogsOpenState[DialogLabels.DELETE_USER]}
-        useAdminListUsersStore={useAdminListUsersStore}
         onClose={() => closeDialog(DialogLabels.DELETE_USER)}
         isDeleting={isDeletingUser}
         handleDeleteUser={handleDeleteUser}
       />
       <EditUserDialog
         open={dialogsOpenState[DialogLabels.EDIT_USER]}
-        useAdminListUsersStore={useAdminListUsersStore}
         isSubmitting={isUpdatingUser}
         onClose={() => closeDialog(DialogLabels.EDIT_USER)}
         handleUpdateUser={handleUpdateUser}
       />
       <AdminDialog
         open={dialogsOpenState[DialogLabels.TOGGLE_ADMIN]}
-        useAdminListUsersStore={useAdminListUsersStore}
         onClose={() => closeDialog(DialogLabels.TOGGLE_ADMIN)}
         isLoading={isUpdatingUserAdmin}
         updateUserAdmin={handleUpdateUserAdmin}
       />
       <ResetPasswordDialog
         open={dialogsOpenState[DialogLabels.RESET_PASSWORD]}
-        useAdminListUsersStore={useAdminListUsersStore}
         onClose={() => closeDialog(DialogLabels.RESET_PASSWORD)}
         handleUpdatePassword={handleUpdatePassword}
       />
