+++ packages/ui/src/views/user-management/components/dialogs/types.ts
@@ -1,0 +1,7 @@
+import { IDialogErrorStates, IDialogHandlers, IDialogLoadingStates } from '@views/user-management/types'
+
+export interface IDialogsProps {
+  handlers: IDialogHandlers
+  loadingStates: IDialogLoadingStates
+  errorStates: IDialogErrorStates
+}
