+++ packages/ui/src/views/user-management/components/index.ts
@@ -1,0 +1,1 @@
+export * from './main'
