+++ packages/ui/src/views/user-management/providers/StoreProvider/StoreProvider.tsx
@@ -1,0 +1,15 @@
+import { createContext } from 'react'
+
+import { StoreContextType, StoreProviderProps } from './types'
+
+export const StoreContext = createContext<StoreContextType | undefined>(undefined)
+
+export const UserManagementStoreProvider = ({
+  children,
+  useAdminListUsersStore,
+  useTranslationStore
+}: StoreProviderProps) => {
+  return (
+    <StoreContext.Provider value={{ useAdminListUsersStore, useTranslationStore }}>{children}</StoreContext.Provider>
+  )
+}
