+++ packages/ui/src/views/user-management/providers/StoreProvider/hooks/useUserManagementStore.ts
@@ -1,0 +1,13 @@
+import { useContext } from 'react'
+
+import { StoreContext } from '@views/user-management/providers/StoreProvider'
+
+export const useUserManagementStore = () => {
+  const context = useContext(StoreContext)
+
+  if (!context) {
+    throw new Error('useUserManagementStore must be used within UserManagementStoreProvider')
+  }
+
+  return context
+}
