+++ packages/ui/src/views/user-management/providers/StoreProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './hooks/useUserManagementStore'
+export * from './StoreProvider'
