+++ packages/ui/src/views/user-management/providers/StoreProvider/types.ts
@@ -1,0 +1,13 @@
+import { TranslationStore } from '@/views'
+import { IAdminListUsersStore } from '@views/user-management/types'
+
+export interface StoreProviderProps {
+  useAdminListUsersStore: () => IAdminListUsersStore
+  useTranslationStore: () => TranslationStore
+  children: React.ReactNode
+}
+
+export interface StoreContextType {
+  useAdminListUsersStore: () => IAdminListUsersStore
+  useTranslationStore: () => TranslationStore
+}
